+++ CMakeLists.txt
@@ -71,10 +71,10 @@ set_package_properties(OpenCV PROPERTIES TYPE OPTIONAL)
 
 if(OpenCV_FOUND)
     set(OpenCV_dep OpenCV)
-    find_package(YARP COMPONENTS os sig math dev cv REQUIRED)
+    find_package(YARP COMPONENTS OS sig math dev cv REQUIRED)
 else()
     set(OpenCV_dep "")
-    find_package(YARP COMPONENTS os sig math dev REQUIRED)
+    find_package(YARP COMPONENTS OS sig math dev REQUIRED)
 endif()
 yarp_configure_external_installation(event-driven)
